planner: move `rule_constant_propagation` to rule pkg. (#55231)
