+++ pkg/planner/core/BUILD.bazel
@@ -67,7 +67,6 @@ go_library(
         "rule_aggregation_skew_rewrite.go",
         "rule_collect_plan_stats.go",
         "rule_column_pruning.go",
-        "rule_constant_propagation.go",
         "rule_decorrelate.go",
         "rule_derive_topn_from_window.go",
         "rule_eliminate_projection.go",
