+++ pkg/planner/core/rule/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "rule",
     srcs = [
         "rule_build_key_info.go",
+        "rule_constant_propagation.go",
         "rule_init.go",
     ],
     importpath = "github.com/pingcap/tidb/pkg/planner/core/rule",
